requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
python-dotenv
openai
anthropic
twilio
pydantic
pydantic-settings
requests
alembic
httpx
